Backend cleanup for question
